+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentMethodType.swift
@@ -19,6 +19,8 @@ extension PaymentSheet {
             switch self {
             case .dynamic("revolut_pay"):
                 return [.returnURL]
+            case .dynamic("amazon_pay"):
+                return [.returnURL]
             case .dynamic("mobilepay"):
                 return [.returnURL]
             case .dynamic("zip"):
@@ -98,6 +100,8 @@ extension PaymentSheet {
                 return "GrabPay"
             } else if case .dynamic("fpx") = self {
                return "FPX"
+            } else if case .dynamic("amazon_pay") = self {
+                return "Amazon Pay"
             } else if case .dynamic(let name) = self {
                 // TODO: We should introduce a display name in our model rather than presenting the payment method type
                 return name